buk cotent table dynmic
