+++ smpp_backend/sms/pom.xml
@@ -96,7 +96,10 @@
 			<artifactId>hibernate-core</artifactId>
 			<version>6.2.6.Final</version>
 		</dependency>
-
+		<dependency>
+			<groupId>org.springframework.boot</groupId>
+			<artifactId>spring-boot-starter-jdbc</artifactId>
+		</dependency>
 		<dependency>
 			<groupId>org.hibernate</groupId>
 			<artifactId>hibernate-entitymanager</artifactId>
