+++ smpp_backend/sms/src/main/java/com/hti/smpp/common/Repository/BulkMgmtContentRepository.java
@@ -7,4 +7,7 @@
 //
 //@Repository
 //public interface BulkMgmtContentRepository extends JpaRepository<BulkMgmtContent, Long> {
+//	
+//	
+//	
 //}
