+++ smpp_backend/sms/src/main/java/com/hti/smpp/common/SmsApplication.java
@@ -7,7 +7,7 @@ import org.springframework.cloud.client.discovery.EnableDiscoveryClient;
 @EnableDiscoveryClient
 @SpringBootApplication
 public class SmsApplication {
-	
+
 	public static void main(String[] args) {
 		SpringApplication.run(SmsApplication.class, args);
 	}
