+++ smpp_backend/sms/src/main/java/com/hti/smpp/common/dto/Json.java
@@ -1,30 +1,14 @@
 package com.hti.smpp.common.dto;
 
-import javax.persistence.Entity;
-import javax.persistence.GeneratedValue;
-import javax.persistence.GenerationType;
-import javax.persistence.Id;
+import org.springframework.stereotype.Component;
 
-@Entity
-public class BulkMgmtContent {
-
-	@Id
-	@GeneratedValue(strategy = GenerationType.AUTO)
-	private Long id;
-
+@Component
+public class Json {
 	private Long destination;
 
 	private String content;
 
 	private String flag = "F";
-
-	public Long getId() {
-		return id;
-	}
-
-	public void setId(Long id) {
-		this.id = id;
-	}
 
 	public Long getDestination() {
 		return destination;
@@ -49,5 +33,5 @@ public class BulkMgmtContent {
 	public void setFlag(String flag) {
 		this.flag = flag;
 	}
-
+	
 }
